pyproject.toml
[project]
name = "jtbd-assistant-platform"
version = "0.1.0"
description = "AI-powered JTBD Assistant Platform using Streamlit, OpenAI, and Supabase"
authors = [
    {name = "JTBD Assistant Team"}
]
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.28.0",
    "openai>=1.3.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "pgvector>=0.2.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "tiktoken>=0.5.0",
    "pydantic>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0"
]
dspy = [
    "dspy-ai>=2.4.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true